+++ CHANGELOG.md
@@ -13,6 +13,8 @@ This project adheres to [Semantic Versioning](http://semver.org/). Additionally,
 
 #### Fixed
 
+*  Enable unit tests for all environment-operator packages. [[BITE-1472](https://agile-jira.pearson.com/browse/BITE-1472)]
+
 ### **[0.0.6] - 2017-09-13 [RELEASED]**
 
 #### Fixed